requirements.txt
monarchmoney
splitwise
python-dotenv